Modify custom asset format to use gob
To accomplish this, a few modifications has been made internally.

The new file format is simply "Ardent", followed by a null byte,
followed by gob-encoded data.
An image is encoded as either a null byte (not an image) or a PNG file.
This is done as PNG's header will never start with a null byte.

Note that the new file format starts with "Ardent" instead of "ardent"
to prevent old and new file format from colliding.

The MarshalBinary and UnmarshalBinary function of asset has been
replaced by Marshal and Unmarshal respectively to make sure gob doesn't
pick them up and use them, causing an endless loop as we do call gob in
those functions now.

The Image field of an asset is now a struct that embeds image.Image.
This allows MarshalBinary and UnmarshalBinary to be implemented on top
of it.

Resolves #15.